Added basic (not fully tweaked) camera system.

+++ Renderer/src/Timer/Timer.h
@@ -15,15 +15,15 @@ public:
         start_time = std::chrono::high_resolution_clock::now();
     }
 
-    double currentTime() const
+    long currentTime() const
     {
-        return std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::high_resolution_clock::now().time_since_epoch()).count();
+        return (long)std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::high_resolution_clock::now().time_since_epoch()).count();
     }
 
-    double elapsed() const
+    long elapsed() const
     {
         auto end_time = std::chrono::high_resolution_clock::now();
-        return std::chrono::duration_cast<std::chrono::milliseconds>(end_time - start_time).count();
+        return (long)std::chrono::duration_cast<std::chrono::milliseconds>(end_time - start_time).count();
     }
 
 private:
